Refactored notifiers
Fixed an error where adding a sample aysnchronously could cause a
NullPointerException which crashed the thread which pulls items off the
queue. This results in the queue filling up and blocking all new
samples.

Also made the aggregating sampler not use the async mechanism as it only
creates a sample once per second.